# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ChangeNow API
# during dev, should point to your local machine
host: localhost:8888
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /flipper/enabled:
    x-swagger-router-controller: flipper
    post:
      description: Check if a feature is enabbled for this user
      operationId: isEnabled
      parameters:
        - name: feature
          in: formData
          description: The name of the feature to retrieve
          required: true
          type: string
        - name: actor
          in: formData
          description: user's login
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/isEnabledResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  isEnabledResponse:
    required:
      - enabled
    properties:
      enabled:
        type: boolean
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string